fix frontend lint errors
